make x,y,w,h size bigger to support bigger display.

+++ tfts/tftframework/common.h
@@ -15,8 +15,8 @@ struct tft_device_data
    void (*send_command)(struct tft_device_data *, uint16_t command);
    void (*send_data)(struct tft_device_data *, uint16_t data);
    void (*init_display)(struct tft_device_data *);
-   void (*set_addr_window)(struct tft_device_data *, uint8_t x0, uint8_t y0,
-			 uint8_t x1, uint8_t y1);
+   void (*set_addr_window)(struct tft_device_data *, uint16_t x0, uint16_t y0,
+			 uint16_t x1, uint16_t y1);
    void (*shutdown_connection)(struct tft_device_data *);
    void (*update_display)(struct tft_device_data *, uint8_t *mem, ssize_t size);
 };
